pyproject.toml
[project]
name = "opssage"
version = "0.1.0"
description = "An AI-powered advisory agent for incident analysis and remediation recommendations in SRE workflows."
authors = [{ name = "Ithaqua Nguyen", email = "ithadev.nguyen@gmail.com" }]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.117.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
    "fastapi>=0.121.1",
    "uvicorn>=0.27.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "python-multipart>=0.0.6",
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "markdown>=3.5.1",
]
requires-python = ">=3.10,<3.14"

[dependency-groups]
dev = [
    "pytest~=8.4.2",
    "pytest-asyncio~=1.2.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
]
deployment = ["absl-py~=2.3.1"]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.18.2",
    "codespell~=2.4.1",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.4.20250913",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["sages"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sages*", "apis*"]
exclude = ["tests*", "docs*", "scripts*", "dashboard*", "docker*", "data*"]
